Update the unit tests
Move the DBus tests to its own directory, update imports and fix the tests.
