+++ tests/nosetests/pyanaconda_tests/__init__.py
@@ -23,15 +23,13 @@ import locale
 
 from textwrap import dedent
 from unittest.mock import Mock, patch
-from xml.etree import ElementTree
 
 from pyanaconda.core.constants import DEFAULT_LANG
-from pyanaconda.dbus.template import InterfaceTemplate
+from dasbus.server.template import BasicInterfaceTemplate
 from pyanaconda.modules.common.constants.interfaces import KICKSTART_MODULE
 from pyanaconda.modules.common.structures.kickstart import KickstartReport
 from pyanaconda.modules.common.task import TaskInterface
-from pyanaconda.dbus.xml import XMLGenerator
-from pyanaconda.dbus.typing import get_native
+from dasbus.typing import get_native
 
 
 # Set the default locale.
@@ -68,24 +66,6 @@ class run_in_glib(object):
             return self._result
 
         return create_loop
-
-
-def compare_xml(test, first_xml, second_xml):
-    """Compare two XML-formatted strings.
-
-    Python 3.8 changed the order of the attributes and introduced
-    the function canonicalize that should be used for testing.
-    """
-    # Prettify the XML.
-    first_xml = XMLGenerator.prettify_xml(first_xml)
-    second_xml = XMLGenerator.prettify_xml(second_xml)
-
-    # Normalize the XML attributes.
-    canonicalize = getattr(ElementTree, "canonicalize", lambda xml, *args, **kwargs: xml)
-    first_xml = canonicalize(first_xml, with_comments=True)
-    second_xml = canonicalize(second_xml, with_comments=True)
-
-    test.assertEqual(first_xml, second_xml)
 
 
 def check_kickstart_interface(test, interface, ks_in, ks_out=None, ks_valid=True, ks_tmp=None):
@@ -175,7 +155,7 @@ def check_task_creation(test, task_path, publisher, task_class):
 
     :param test: instance of TestCase
     :param task_path: DBus path of the task
-    :param publisher: Mock instance of the pyanaconda.dbus.DBus.publish_object
+    :param publisher: Mock instance of the publish_object method
     :param task_class: class of the tested task
 
     :return: instance of the task
@@ -192,7 +172,7 @@ def check_task_creation_list(test, task_paths, publisher, task_classes):
     :param test: instance of TestCase
     :param task_paths: DBus paths of the tasks
     :type task_paths: [str]
-    :param publisher: Mock instance of the pyanaconda.dbus.DBus.publish_object
+    :param publisher: Mock instance of the publish_object method
     :param task_classes: list of classes of the tested tasks; the order is important here
 
     :return: list of instances of tasks
@@ -210,7 +190,7 @@ def check_dbus_object_creation(test, path, publisher, klass):
 
     :param test: instance of TestCase
     :param task: DBus path of the published object
-    :param publisher: Mock instance of the pyanaconda.dbus.DBus.publish_object
+    :param publisher: Mock instance of the publish_object method
     :param klass: class of the tested DBus object
     """
     publisher.assert_called_once()
@@ -218,7 +198,7 @@ def check_dbus_object_creation(test, path, publisher, klass):
 
     test.assertEqual(path, object_path)
     test.assertIsInstance(obj.implementation, klass)
-    test.assertIsInstance(obj, InterfaceTemplate)
+    test.assertIsInstance(obj, BasicInterfaceTemplate)
 
     return obj
 
@@ -230,7 +210,7 @@ def patch_dbus_publish_object(func):
 
     # TODO: Extend this to patch the whole DBus object and pass in a useful abstraction.
     """
-    return patch('pyanaconda.dbus.DBus.publish_object')(func)
+    return patch('pyanaconda.core.dbus.DBus.publish_object')(func)
 
 
 def patch_dbus_get_proxy(func):
@@ -238,4 +218,4 @@ def patch_dbus_get_proxy(func):
 
     This is a shortcut to avoid creating of DBus proxies using DBus.
     """
-    return patch('pyanaconda.dbus.DBus.get_proxy')(func)
+    return patch('pyanaconda.core.dbus.DBus.get_proxy')(func)
